Add abi break checker
This program compares the exported symbols in all the libraries
to detect if two versions are binary compatible.

It is still a work-in-progress, since there are several problems,
like some binaries exporting symbols that aren't in the headers
due to functions not marked as "static". Some work is required in
this field.