Allow to specify versions to ignore
Sometimes, there are new versions that fail to build, or contain
important bugs, and it's not desirable to update to them. For
these cases, this new parameter allows to specify a list of
versions to ignore.

Example: Cairo version 1.17.8 fails to build in Gnome SDK, so
it is ignored with:

    # ext:updatesnap
    #   version-format:
    #     ignore-version: '1.17.8'